Added check for  StoreManagerAdUsers

+++ TervisRemoteDesktopServices.psm1
@@ -83,8 +83,12 @@ function Update-StoreManagerToStoresRdsPrivilege {
     }
     $GroupMembers = Get-ADGroupMember -Identity 'Privilege_StoresRDS_RemoteDesktop'
     foreach ($GroupMember in $GroupMembers) {
-        If (-NOT (($GroupMember).DistinguishedName -like "*OU=Store Accounts,*" -or ($GroupMember).DistinguishedName -in ($StoreManagerAdUsers).DistinguishedName)) {
-            Remove-ADGroupMember -Identity 'Privilege_StoresRDS_RemoteDesktop' -Members ($GroupMember).DistinguishedName -Confirm:$false
+        If ($StoreManagerAdUsers) {
+            If (-NOT (($GroupMember).DistinguishedName -like "*OU=Store Accounts,*" -or ($GroupMember).DistinguishedName -in ($StoreManagerAdUsers).DistinguishedName)) {
+                Remove-ADGroupMember -Identity 'Privilege_StoresRDS_RemoteDesktop' -Members ($GroupMember).DistinguishedName -Confirm:$false
+            }
+        } else {
+            Throw "The StoreManagerAdUser Variable is empty"
         }
     }
 }
